commit confirm and login fix
core.py:
fixed logging in as root for op_commands to ensure _session exists
first.
commit_confirm parameter changed to confirmed to be in line with
ncclient,
also moved commands to be the first parameter of the commit() function,
since it is the most common.

jaide_cli.py:
changed commit_confirm to be confirmed to be in line with ncclient.
changed argument handling for —confirmed to be seconds instead of
minutes, since we aren’t on our custom ncclient version any more.